# backend/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "College Event Management API",
    "description": "API for managing college events, registrations, and notifications",
    "version": "1.0.0"
  },
  "host": "localhost:5000",
  "basePath": "/api",
  "schemes": ["http"],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT Authorization header using the Bearer scheme"
    }
  },
  "paths": {
    "/auth/register": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Register a new user",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": { "type": "string" },
                "email": { "type": "string" },
                "password": { "type": "string" },
                "role": { "type": "string", "enum": ["student", "organizer", "admin"] }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "User registered successfully"
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Login user",
        "parameters": [
          {
            "name": "credentials",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": { "type": "string" },
                "password": { "type": "string" }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Login successful",
            "schema": {
              "type": "object",
              "properties": {
                "token": { "type": "string" },
                "user": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string" },
                    "name": { "type": "string" },
                    "email": { "type": "string" },
                    "role": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/events": {
      "get": {
        "tags": ["Events"],
        "summary": "Get all events",
        "responses": {
          "200": {
            "description": "List of events"
          }
        }
      },
      "post": {
        "tags": ["Events"],
        "summary": "Create a new event",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "event",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "title": { "type": "string" },
                "description": { "type": "string" },
                "date_time": { "type": "string", "format": "date-time" },
                "location": { "type": "string" },
                "capacity": { "type": "integer" },
                "category": { "type": "string" }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Event created successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/events/{id}": {
      "get": {
        "tags": ["Events"],
        "summary": "Get event by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Event details"
          },
          "404": {
            "description": "Event not found"
          }
        }
      }
    },
    "/events/upcoming": {
      "get": {
        "tags": ["Events"],
        "summary": "Get upcoming events",
        "responses": {
          "200": {
            "description": "List of upcoming events"
          }
        }
      }
    },
    "/registrations/{eventId}": {
      "post": {
        "tags": ["Registrations"],
        "summary": "Register for an event",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Registration successful"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Event not found"
          }
        }
      }
    },
    "/registrations/my": {
      "get": {
        "tags": ["Registrations"],
        "summary": "Get user's registrations",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of user's registrations"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/registrations/check/{eventId}": {
      "get": {
        "tags": ["Registrations"],
        "summary": "Check if user is registered for an event",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Registration status",
            "schema": {
              "type": "object",
              "properties": {
                "isRegistered": { "type": "boolean" }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/notifications": {
      "get": {
        "tags": ["Notifications"],
        "summary": "Get user's notifications",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of notifications"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    }
  }
} 